[APPC-1437] Moved files to correct package
