+++ app/src/main/java/com/asfoundation/wallet/repository/AutoUpdateRepository.kt
@@ -1,12 +1,13 @@
 package com.asfoundation.wallet.repository
 
-import com.asfoundation.wallet.interact.AutoUpdateModel
-import com.asfoundation.wallet.interact.AutoUpdateService
+import com.asfoundation.wallet.service.AutoUpdateService
+import com.asfoundation.wallet.viewmodel.AutoUpdateModel
 import io.reactivex.Single
 
 class AutoUpdateRepository(private val autoUpdateService: AutoUpdateService) {
 
-  private var autoUpdateModel: AutoUpdateModel = AutoUpdateModel()
+  private var autoUpdateModel: AutoUpdateModel =
+      AutoUpdateModel()
 
   fun loadAutoUpdateModel(invalidateCache: Boolean): Single<AutoUpdateModel> {
     if (autoUpdateModel.isValid() && !invalidateCache) {
