+++ app/src/main/java/com/asfoundation/wallet/service/AutoUpdateService.kt
@@ -1,5 +1,7 @@
-package com.asfoundation.wallet.interact
+package com.asfoundation.wallet.service
 
+import com.asfoundation.wallet.entity.AutoUpdateResponse
+import com.asfoundation.wallet.viewmodel.AutoUpdateModel
 import io.reactivex.Single
 import retrofit2.http.GET
 
@@ -8,7 +10,8 @@ class AutoUpdateService(private val api: AutoUpdateApi) {
   fun loadAutoUpdateModel(): Single<AutoUpdateModel> {
     return api.getAutoUpdateInfo()
         .map {
-          AutoUpdateModel(it.softUpdate.versionCode, it.softUpdate.minSdk,
+          AutoUpdateModel(it.softUpdate.versionCode,
+              it.softUpdate.minSdk,
               it.blackList)
         }
         .onErrorReturn { AutoUpdateModel() }
